Disable mouse in tmux
I tried this for a while and I decided I don't like it:
- if you highlight text with the mouse, it isn't copied to the system
clipboard, this is massively annoying.
- clicking and scrolling with the mouse just isn't that useful since the
same things can all be accomplished faster with good vi key mappings and
so in practice I almost never used it, but I do copy stuff with the
mouse fairly regularly...
- on laptops you may accidentally click or scroll if palm rejection
isn't working well, this hasn't been a problem lately, but it was in the
past on the 9343. Newer debian or gnome or libinput must have improved
the palm rejection, but it definitely was an issue. It's probably
hardware/driver dependent, so I wouldn't be surprised to see the issue
come back.