Syntastic: don't auto-open buffer with issue list
Older scripts have a lot of issues that won't be fixed in this
particular edit, and Syntastic/shellcheck will add a purple arrow in the
sidebar for each issue so you can still see issues for lines your
changing. The buffer list just takes up 1/3 of the screen for issues you
probably aren't going to fix this session anyway.