dotfiles/nvim/.config/init.vim: simplify nvim conf
Nvim has grown a lot of new features like native LSP support, so remove
the plugins that support that. Similarly we shouldn't need syntastic,
racer, and deoplete anymore with LSP. Also remove all the config for
those plugins.

We don't write go code or bitbake code anymore, so remove the plugins
for those.

The install for the `fzf` plugin changed slightly, so update that.

I don't think we need the python support anymore, so remove that.