dotfiles: restructure for windows
Now that we have dploy for windows, make a new dotfiles script
`dotfiles_win.sh` and put all the win stuff there. The other one remains
linux only. We also need a .dploystowignore, same as .stowignore. We
don't need to stow all the same files that we did on linux.