All scripts: force the user to set HOME
On windows, if the environment variable HOME does not equal USERPROFILE,
it means that HOMEDRIVE/HOMESHARE and HOMEPATH are set to something
other than the local disk (like a network share in a corporate setting).
Then git bash's env variable HOME is set to HOMEDRIVE+HOMEPATH on that
share and everything breaks.

Before we were scripting around this by symlinking stuff into
USERPROFILE, but then we had to fix bash_profile to reset that
environment var or every other tool would break. But we had to do that
on the network share also, since bash is going to look in ~/. We can't
symlink (or at least its a bad idea) anything across the network share,
so we would have to use a cp. This whole mess was very brittle.

So we change it to:
- check when the script starts, if we are on windows, if HOME and
  USERPROFILE are the same, if they are life is good
- if they are not the same, tell the user they must fix that first and
  exit.
- Tell the user how to fix it.