.gitconfig: autocrlf = input is a bad idea
See
https://stackoverflow.com/questions/1967370/git-replacing-lf-with-crlf,
better idea is to just make the editor (in our case generally nvim) save
with the correct line endings in the first place and disable gits
attempt at being smart. Any other editor that does the wrong thing won't
be used very often and the output can be fixed with the dos2unix
program.